requirements.txt
flask==3.0.0
slack-sdk==3.33.4
openai==1.54.4
requests==2.32.3
python-dotenv==1.0.1
Pillow==10.4.0
gunicorn==23.0.0
httpx==0.27.2
stripe==7.8.0
supabase==2.16.0
# Async and performance dependencies
aiohttp==3.9.1
asyncio-throttle==1.0.2
redis==5.0.1
rq==1.16.0
asyncpg==0.29.0
uvicorn==0.24.0
fastapi==0.104.1
pydantic==2.5.0
# Monitoring and caching
prometheus-client==0.19.0
structlog==23.2.0 